requirements.txt
django==1.10
djangorestframework==3.5
markdown==2.6.7
addict==2.0.0